Refresh the Repository Store on webhook events and startup for all installation instances. (#84)
Previously, the application listened to GitHub webhook events and added/removed individual items from a repository_store. However, when filtering by topic it made a call to the GitHub API. This is because the GitHub webhook event data does not include topic values for repositories.

This update makes the application completely refresh the repository_store for all webhook events and startup. This loads the appropriate topic data. Filtering is now based on the repository store, with no requirement to call the GitHub API.

* Rewrote index and repositories tests; removed initialization features
* Making the code more robust
* Added listener for repository events
* Simplify routing so that /tako is no longer required.
* Simplify routing so that /tako is no longer required.
* Subscribe to repository webhook events
* Tako now supports multiple installations
* Updated test to cover multiple installations
* Apply suggestions from code review
* Output more repo properties; Updated test for repositories; Reverted routing to include /tako